Telepathist (Artificial Intelligence Project)
To replicate the idea of  20 questions game where a user is asked to
think of an animal the program guesses the answer on the basis of  yes
or no answers given by the user to the questions asked by the program.
If the data is not in the database then the program asks the user to add
the animal name and asks some specific questions about it. It works on
the principle of  Artificial Intelligence and Machine Learning . The
project was implemented in python for both backend connectivity and GUI
and the database used was Sqlite 3.